src/requirements.txt
snowflake-connector-python
snowflake-snowpark-python
pandas
numpy
datetime
fastapi
